카테고리 관리자 권한 인증

See merge request sw_track/class_02_seoul/web_project/team7/shopping-mall-racer!14

+++ src/services/category-service.js
@@ -1,11 +1,12 @@
-const { categoryModel } = require("../db");
+const mongoose = require("mongoose");
+const { categoryModel, userModel } = require("../db");
 
 class CategoryService {
-  constructor(categoryModel) {
+  constructor(categoryModel, userModel) {
     this.categoryModel = categoryModel;
+    this.userModel = userModel;
   }
 
-  
   async getAllCategory() {
     try {
       const category = await this.categoryModel.findAll();
@@ -25,12 +26,13 @@ class CategoryService {
     }
   }
 
-  async insertCategory(name) {
+  async insertCategory(name, userId) {
     try {
       const isExist = await this.categoryModel.findByName(name);
       if (isExist) {
         throw new Error("name is already exist.");
       }
+      await this.checkIsAdministrator(userId);
 
       const result = this.categoryModel.create(name);
       return result;
@@ -40,8 +42,10 @@ class CategoryService {
     }
   }
 
-  async updateCategory(currentCategoryName, nameToChange) {
+  async updateCategory(currentCategoryName, nameToChange, userId) {
     try {
+      this.checkIsAdministrator(userId);
+
       const result = await this.categoryModel.update(
         currentCategoryName,
         nameToChange
@@ -53,8 +57,9 @@ class CategoryService {
     }
   }
 
-  async deleteCategory(name) {
+  async deleteCategory(name, userId) {
     try {
+      this.checkIsAdministrator(userId);
       await this.categoryModel.delete(name);
       return;
     } catch (error) {
@@ -62,8 +67,16 @@ class CategoryService {
       next(error);
     }
   }
+
+  async checkIsAdministrator(userId) {
+    const ObjectId = require("mongodb").ObjectId;
+    const user = await this.userModel.findById(ObjectId(userId));
+    if (user.role !== "admin") {
+      throw new Error("Request is not allowed. The user is not administrator.");
+    }
+  }
 }
 
-const categoryService = new CategoryService(categoryModel);
+const categoryService = new CategoryService(categoryModel, userModel);
 
 export { categoryService };
